+++ Classes and objects/Call a method from another method/call_method.py
