+++ Classes and objects/Call a method from another method/tests/test_task.py
@@ -1,6 +1,6 @@
 import unittest
 
-from task import Calculator
+from call_method import Calculator
 
 
 class TestCase(unittest.TestCase):
