+++ Classes and objects/Class and Instance Variables 2/class_instance2.py
