+++ File input output/Read all lines/read_all_lines.py
