+++ Loops/Else with loops 2/tests/test_task.py
@@ -3,7 +3,7 @@ import contextlib
 import io
 
 
-from task import contains_even_number
+from else_with_loops2 import contains_even_number
 
 
 class TestCase(unittest.TestCase):
