+++ Loops/Fix infinite execution/infinite.py
