+++ Modules and packages/Executing modules as scripts/task-info.yaml
@@ -7,8 +7,9 @@ files:
     length: 18
     placeholder_text: '# Call func() from the imported module'
   - offset: 107
-    length: 100
-    placeholder_text: print('This should be printed ONLY when task.py is called directly.')
+    length: 105
+    placeholder_text: print('This should be printed ONLY when dict_keys.py is called
+      directly.')
 - name: tests/test_task.py
   visible: false
 - name: __init__.py
