+++ Variables/Comparison operators/task-info.yaml
@@ -9,16 +9,16 @@ files:
   - offset: 367
     length: 1
     placeholder_text: '?'
-  - offset: 522
+  - offset: 519
     length: 1
     placeholder_text: '?'
-  - offset: 530
+  - offset: 527
     length: 1
     placeholder_text: '?'
-  - offset: 711
+  - offset: 708
     length: 2
     placeholder_text: '?'
-  - offset: 863
+  - offset: 860
     length: 2
     placeholder_text: '?'
 - name: tests/test_task.py
